install rabbitmq-server in serial
In order to prevent race conditions with nodes joining the cluster
simultaneously when the cluster is first formed, we move the rabbitmq
installation play to be 'serial: 1'. However, when the nodes are being
upgraded, it cannot be done in serial so in this case we set 'serial: 0'

The tests are removed from a post_task include in the install play, and
moved to their own play as they need to be run after the entire cluster
has been formed. As well as moving a few generic vars into the
test-vars.yml include, we also pass in the specific version of rabbitmq
to be tested against in the test play.

Fixes-Bug: #1573030

Change-Id: Id119ff9f20ddfd8e1f29598c8c5ce862d2e7fab4
(cherry picked from commit 5dc67955f0ac08a7c9719d641e9828558620da89)