Fix get_url SNI issues in CentOS 7
CentOS 7 comes with Python 2.7.5, which does not have support for
servers that use Server Name Indication (SNI) to host more than
one SSL certificate on the same IP address. Three Python modules
are required on CentOS 7 systems to add this support:

* pyasn1
* pyOpenSSL
* ndg-httpsclient

This patch should prevent the SSL verification issues that occur
when the RabbitMQ RPMs are downloaded with get_url.

Closes-bug: 1604922

Change-Id: Ia14f25ed0ae399cfea6177adb6d34582a40c9249