Add repo installation method
In some cases, deployers may want to use a repo instead of a deb
file (limited connectivity, repo mgmt  tools already in place, ...)

This commit introduces this behavior in an optional manner: The
default behavior is still using a deb file. If using an external
repository, the repo used will be RabbitMQ recommended one hosted
on packagecloud.io.

Change-Id: If366d7411cffd59e4988ceba6fc47eac2b384ddf