Pin erlang packages to version 19.3
Erlang 20 has been released and is not compatible with the current
release of RabbitMQ.[0] Pin erlang packages to the latest version
version supported by RabbitMQ 3.6.x, 19.3.

For package pinning to work correctly on Ubuntu, the priority of pinning
to a specific version needs to be higher than the priority of all
packages coming from the ESL repo.

For RedHat distros, the yum-versionlock plugin will need to be installed
and used to properly pin a package to a specific version.
[0] https://groups.google.com/forum/#!topic/rabbitmq-users/_imbAavBYjY

Combined back port of the following commits:
2a3091c9c2ad38376c15d740fa90488c5790cddb
c73e469e197ffb856f8117786061d79a830918b6

Change-Id: Ia8f053e499db7aa7c78c60213d3154d967b92a98