Actually pin erlang packages
For package pinning to work correctly on Ubuntu, the priority of pinning
to a specific version needs to be higher than the priority of all
packages coming from the ESL repo.

For RedHat distros, the yum-versionlock plugin will need to be installed
and used to properly pin a package to a specific version.

Change-Id: Ie0c241de6a1a371969ed6e2886d6514c9419622f