Add final dnf support bits
The dnf package manager was missing from the list of package
managers in the upgrade check tasks. This patch ensures that
the upgrade check runs on servers with dnf.

The dnf module has no 'update_cache' argument, so this patch removes
that if dnf is being used. Updating the metadata cache is not needed.

Implements: blueprint centos-and-dnf
Change-Id: I10565b2e9379d5ac60ae65bfcdda221a0218988e