tests: test-vars: Install the RabbitMQ file directly
The distribution repositories may provide different RabbitMQ versions
and even change during time. As such, in order to be able to reliably
test the upgrade path, we need to install the rpm/deb packages directly
instead of replying to the distribution repositories. This is mostly a
problem for openSUSE which uses the OBS repositories since there is not
an upstream one for it.

Change-Id: I3641da0cf67809ca0b343a9692fd5c7592fadb78