Create cluster without using a failed task
Currently a new cluster is created by using a try/rescue
operation which always fails for a new cluster, causing
a lot of confusion due to the failed task report.

This patch implements an improvement which tests the
cluster state and creates a new one if it is not ready
without requiring the task to fail.

As the cluster creation is a single task, it's merged
into the rabbitmq_cluster.yml task file.

As the tags are all the same, the tag is set in the
main task file instead of individually on all the tasks.

Change-Id: Iaa8e429cbd1c539d8242453c0490373c758dd9fe