Check and set cookie without using a failed task
Currently the checking of the cookie content uses
a try/rescue block and for a new install always
results in a failed task. This causes a lot of
confusion due to the failed task report.

This patch improves the process by executing the
stop/start if the cookie file is changed in any
way by the copy task. This is far simpler and
results in no task fails.

As all the tasks related to this were marked
rabbitmq-config the tag implementation is set
in the main tasks file and removed from all the
individual tasks.

Change-Id: Iba959f730f2453bf9a295158749b96ed3a966ef4