Join cluster without using a failed task
Currently a new node joining the cluster is done by using
a try/rescue operation which always fails for a new node,
causing a lot of confusion due to the failed task report.

This patch ensures that the implementation no longer does
that.

In order to prevent lint check failures, each join task
has a 'changed_when: true' added.

Change-Id: Ic1da9f3ad1016831fe37643165880e7ff98ca923