Implement separate deploy and upgrade tests
The current functional test does a full upgrade process,
but uses the same current role for both the previous
version and the new version.

Ideally we want to test the initial deploy using the
previous series' role, so that we're actually testing
a series upgrade, not just a rabbitmq upgrade.

This patch separates the upgrade test out and makes the
default test just be a deployment. This allows us to
more clearly see where something breaks.

Depends-On: https://review.openstack.org/543500
Depends-On: https://review.openstack.org/543236
Change-Id: I43e1e224cd4237ac37b40fde5ce7c279331a012d