Explicitly ensure when apt cache updates happen
In https://review.openstack.org/547015 we adjusted
the apt installation process to ensure that the cache
update happens after the repository changes, but left
out the explicit instruction not to do it in the task
which changes the repo config.

This patch adds that, but also seperates out the cache
update into its own task so that the package install
task remains more readable.

Partial-Bug: #1750656
Change-Id: Id70ad28caa63267c2bb17d8c56a8fa6e00cf55d4