Update rabbitmq to the latest stable release
This change updates our version of RabbitMQ server and Erlang to the
latest stable releases. This change adds the distro file for ubuntu
18.04 and 16.04 which replaces the distro family file "debian". This is
being done because the archives for the two distros are different. While
the SHA256 sum for the packges is the same today, it stands to reason
that they will differ in the near future.

Change-Id: I9784aa10becbb9dbcfc9222c80a7224201d67741
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>