RedHat: Use the distribution package
The RDO repository provids a rabbitmq-server which is tested by the
RDO team so we should use that instead of the external one. We also
skip several tasks which add gpg keys and external repositories
since none of these steps is necessary when we install packages
from the distro repos.

Change-Id: I7c0a988f3aab14042ac2491b1021e789ba9ef4b3