Use vendored RPM GPG keys for RabbitMQ
It seems that the URLs which are being used to deploy the GPG
keys from are not reliable and they could possibly change, this
patch moves them within the repository to make sure that they
are always accessible.

This is a re-implementation rather than a backport - the Stein
(aka master) branch only uses the 'distro' install method, so
this code path is not exercised.

Change-Id: I28199ce149f6893d688d11177ec950b17dbf0886
(cherry picked from commit deccf93bdda1aa873b956418168368284509c99b)