Ensure that RedHat/SUSE only use a distro install
Previously the 'distro' install method was implemented as the
default for CentOS and SUSE, however, some remnants were left
behind, making the configuration confusing and causing some
issues due to the gpg verification on RedHat trying to use the
RabbitMQ gpg check.

Rather than try and keep the many confusing installation methods
for these distributions, we simply remove the other options and
all the config entries related to it.

As part of this, we need to clean up the functional tests which
previously implemented checks based on the 'file' install type
which is no longer used by any distribution.

Change-Id: I28199ce149f6893d688d11177ec950b17dbf0886