Update and level set ubuntu installs
The rabbitmq package installed on both xenial and bionic is actually
the same, allowing the vars files to be consolidated. The package has
also been updated to the latest stable release, 3.7.8.

The repo used for distro based installs has also been changed to
packagecloud.io. New packages have not been published to rabbitmq.com
for a while, the latest release there was 3.6.15.

Change-Id: Id49a8788593c5d64fc9c149c8912565bc3b05a3f