Use in-repo GPG keys
We make remote network hits to get the GPG keys which are quite
unreliable, and apt_key does not support using a proxy properly [1]
so let's store them inside the role and use them.

The implementation here matches that which was done in the
galera_client role in I520ccbadf3320b0d07fc83e3dbec9ea2bd16ec83

This is a re-implementation rather than a backport - the Stein
(aka master) branch only uses the 'distro' install method, so
this code path is not exercised.

Also note that the Erlang yum gpg key and the rabbitmq yum gpg key
are the same, and the Erlang key was never imported - so we've
removed it.

[1] https://github.com/ansible/ansible/issues/31691

Closes-Bug: 1810533
Change-Id: I2715c904975b7940af72bd422904e748d3bae953
(cherry picked from commit 83affc627fd2132bb5a65c4b1f5a07a9f95a7998)