Improve the Highly Available (Mirrored) Queues policy
Without this policy, the queues reply_*, *_fanout_* are mirrored, across
all the RabbitMQ nodes.
It is not necessary to mirror the queues above, because they are usually
bound to one specific consumer, so the mirroring is not essential.
The new policy reduces the number of mirror queues and increases, in
general, the performances because it reduces the number of copies across
the cluster.

More info about RabbitMQ HA [1]

[1]: https://www.rabbitmq.com/ha.html

ref:
http://eavesdrop.openstack.org/meetings/oslo/2019/oslo.2019-02-25-15.22.txt

Change-Id: I783d314aaa68b09abb4fed90818165b9e61e9758