el7: switch to external_repo install_method
This patch changes the default deployment back to using the
RabbitMQ external repositories.  This restores the behaviour
to the same as the old and ships the 3.7.x series.

We deploy without HiPE by default however our testing is doing
it with HiPE.  This patch drops it so we can actually test
behaviour that we deploy by default.

Change-Id: I49747e104f118ef1fb2bdeb329c92d2f5fb66c56