Fix service restart when using tags
If an adjustment to the rabbit config is made which rewrites the
config file, and the playbook is run with
--tags rabbitmq-config,rabbitmq_server-config then the new config
file will be written but the service is not restarted.

This patch adds tag inheritance to the include_tasks for service
restarts to ensure that the service is restarted if the just the
config is updated.

Change-Id: Ia4bcc7a9421d16a8bc35340e1a31a3eb15519369