Fix erlang upgrades for rabbitmq external_repo installs
Upgrading only the 'erlang' package does not pull in updated
dependencies which can cause rabbitmq to fail to start during an
upgrade.

The external and distro dependencies appear to match, so these
are conformed to ensure that all packages are updated during an
upgrade.

This patch additionally switches deployments to the 'erlang-base'
rather than 'erlang-base-hipe' package which appears to have been
discontinued.

Change-Id: I8bd010cba13946c9c46c3792624841e41e3c2eaf