Bump rabbitmq version back
With commit [1] rabbitmq version has been bumped quite extremely
so that with upgrade to V or even W there will be downgrade of rabbitmq
which is really best to avoid.

Mentioned patch was released under EM and was never released. So it's
better to bump rabbit version down to the version V and W have.

[1] https://review.opendev.org/c/openstack/openstack-ansible-rabbitmq_server/+/850350

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/853029/
Change-Id: I5b17878d497504f7bd97b2dccd72ecc28484b249