Do not use 'always' tag in inappropriate places
Using 'always' tag in combination with condition depending on
variables is not always a good idea because these variables may not be
defined at the moment.

Currently, rabbitmq_server role cannot be executed with tag limit,
because task containing 'always' tag has a condition operating on
variable(s) that are registered only when 'rabbitmq-config' tag is being
used.

Change-Id: I7ef7b3fed79933b4c8bbc39a63ca6db54532e10e