Use el/9 repo urls instead of el/8 since a repo is available for it
As we target el9 and not el8, we should use the el/9 repo as it is
available, even though it provides RPMs with an `el8` dist tag. This can
help prevent confusion for users who see a repo on the system configured
with `el/8` in the repourl.

There should be no impact on existing installations, as the same repo
content is provided in both repos in almost all cases--namely, a few
early versions of erlang are unavailable, but as we pin to a version
that exists in both repos, there is no impact to this change: at worst,
the erlang package will be upgraded to a newer minor version if a user
somehow has an older one installed.

Change-Id: I70c13f7f882ce8666b12314a31f76fe63599677c