Remove RabbitMQ restart when changing policy
My assumption is that this restart only exists for historical
reasons or in order to force client re-connection and
re-definition of queues to match this policy.

Unfortunately the current task only executes on a single RabbitMQ
node when part of a cluster, so it is ineffective in forcing
reconnection. From a brief review at least when shifting from
non-HA to HA this also isn't required as queues are immediately
updated to reflect the new policy.

Given no clear reason for the task to be present, I'm removing
it ahead of making changes to the default HA policy.

Change-Id: I956abe68d4d7921a4e87ae5a7694d49546bdde94
(cherry picked from commit cfb1c281510700d507806cb54f9f7843a1f082c1)