Fail in when installing RabbitMQ/Erlang versions not possible
Current behaviour with just pinning on Ubuntu/Debian leads to potential
situations, when there's no selected version available in repo to
proceed with what's available based on set priorities.

Instead, it might be worth to fail explicitly when requested version is
not available. This also aligns behaviour with EL distros.

Change-Id: Iad69cc4b5f8ca40246bcbc604e77eef7975eb104