Move RabbitMQ restart to handlers
Due to historical reasons RabbitMQ operations, like start/stop/restart
were perfromed in tasks and triggered on changes.

As of today there seems to be no need in such approach and we can use
handlers instead.

It will also move some tasks, which intend to run when service is
running, to their own file. This also allows to run them once on cluster
rather then for each host.

We also merge set_cookie with post_install as there is no reason to
maintain the task in it's own file.

Change-Id: I222b9de3ca531724b05c7fa2cf032a3271055d64