Ensure that rabbitmq-erlang and rabbitmq-server repos use the same pins
There is currently a mismatch between the apt pin priorities for the
erlang and server rabbitmq repos. This patch sets the pinning to be
the same.

This should not affect normal package installation as recent changes
to this role require installation of a specific version and will
fail "safe" should that be unavailable. The intention with this patch
is to ensure that the behaviour is the same for both the erlang
and server repos in an abnormal situation where the installation
might fail.

Change-Id: I0d988572a102ca148059d198379d2a3115ecb795