Service must always be restarted during an upgrade
The service is unconditionally stopped during the upgrade
prep tasks. There are several conditions that will notify
the handler to restart the service, such as updating the
config file or installing a new version of the rabbitmq
package. It is possible that none of these things happen
resulting in the service not being restarted during an upgrade.

This patch ensures that the service is always restarted
during an upgrade by notifying the restart handler when
the service is sucessfully stopped in the upgrade prep tasks.

Change-Id: I9fbe7806b3531921b90ac91c17b6a07deca2a4d2
(cherry picked from commit b09ac39cbca11f7a5a14731e583246fe6a6c420e)