Fail with human-readable errors if upgrade impossible
At the moment we simply skip the upgrade preparation tasks
in case cluster is not healthy or meeting expectations for the role
to proceed with the upgrade. This may lead to unexpected behaviour
and unobvious failures later on in the role.

Instead, we add another failure before node upgrade preparation is
run which will display all noticed inconsistencies and human readable
message for each occurance.

Change-Id: I0f252d08b2bcb94bca142dce5e376b5c3bc06417