Fix quorum/stream queues if they're below minimal size
In case it is unsafe to  shutdown current node due to queues going
below their minimal quorum size, we attempt to fix them by growing
affected queues to rest of the cluster.

This step is vital one in case of OS upgrades, when operating system
on controller is being re-installed, as quorum/stream queues
must be manually stretched after re-install.

Change-Id: I87e2ae7b43dbcf77c9f166656686d25dbd8937fd